+++ src/app.h
@@ -1,16 +1,20 @@
 #pragma once
 
 struct app_state {
-    bool show_functions = true;
-    bool show_disasm = true;
-    bool show_hex = true;
-    bool show_strings = true;
-    bool show_log = true;
+    bool show_left = true;
+    bool show_view = true;
+    bool show_right = true;
+    bool show_bottom = true;
     bool show_regs = true;
     bool show_demo = false;
+    bool graph_mode = false;
     char filter[128] = {};
+    char jump_buf[32] = {};
     int selected_func = 1;
     int selected_line = 4;
+    int left_tab = 0;
+    int right_tab = 0;
+    int bottom_tab = 0;
     char log_buf[8192] = {};
     int log_len = 0;
     char file_name[256] = "sample.exe";